+++ Tests/SentryObjCTests/SentryObjCFeedbackApiTests.m
@@ -40,11 +40,11 @@
                  }];
 }
 
-- (void)testSetShakeGestureEnabled_whenFeedbackNotConfigured_shouldNotCrash
+- (void)testFeedbackOnShake_whenFeedbackNotConfigured_shouldNotCrash
 {
     // -- Act & Assert (no crash) --
-    [SentryObjCSDK.feedback setShakeGestureEnabled:YES];
-    [SentryObjCSDK.feedback setShakeGestureEnabled:NO];
+    [SentryObjCSDK.feedback enableFeedbackOnShake];
+    [SentryObjCSDK.feedback disableFeedbackOnShake];
 }
 
 @end